move classes to better package structure
