+++ app/src/main/AndroidManifest.xml
@@ -5,6 +5,7 @@
     <uses-permission android:name="android.permission.INTERNET" />
 
     <application
+        android:name=".BudgetBeeApplication"
         android:allowBackup="true"
         android:fullBackupContent="true"
         android:icon="@mipmap/ic_budget_bee_logo"
@@ -13,31 +14,23 @@
         android:supportsRtl="true"
         android:theme="@style/BudgetBeeMainTheme">
 
-        <activity android:name=".ui.AddTransactionActivity"></activity>
+        <activity android:name=".ui.transaction.AddTransactionActivity" />
 
         <activity
-            android:name=".ui.AddTransactionCategoryActivity"
+            android:name=".ui.transaction.transactions_category.AddTransactionCategoryActivity"
             android:label="@string/title_activity_add_transaction_category"
-            android:theme="@style/BudgetBeeMainTheme.NoActionBar"></activity>
+            android:theme="@style/BudgetBeeMainTheme.NoActionBar" />
 
-        <activity android:name=".ui.TransactionCategoryActivity"></activity>
+        <activity android:name=".ui.transaction.transactions_category.TransactionCategoryActivity" />
 
         <activity
             android:name=".ui.LandingActivity"
             android:theme="@style/BudgetBeeMainTheme">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
+                <action android:name="android.intent.action.VIEW" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
-        </activity>
-
-        <activity
-            android:name=".ui.Registration"
-            android:label="@string/title_activity_registration"
-            android:theme="@style/BudgetBeeMainTheme">
-            <intent-filter>
-                <action android:name="android.intent.action.VIEW" />
             </intent-filter>
         </activity>
     </application>
