+++ app/src/main/java/za/co/app/budgetbee/data/model/database/TransactionCategoryDataModel.kt
@@ -1,4 +1,4 @@
-package za.co.app.budgetbee.data.model
+package za.co.app.budgetbee.data.model.database
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
