+++ app/src/main/res/layout/activity_add_transaction_category.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:layout_margin="@dimen/gutterSpace"
     android:orientation="vertical"
-    tools:context=".ui.AddTransactionCategoryActivity">
+    tools:context=".ui.transaction.transactions_category.AddTransactionCategoryActivity">
 
     <com.google.android.material.textfield.TextInputLayout
         android:layout_width="match_parent"
