+++ app/src/main/res/layout/activity_transaction_category.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.TransactionCategoryActivity">
+    tools:context=".ui.transaction.transactions_category.TransactionCategoryActivity">
 
     <androidx.viewpager.widget.ViewPager
         android:id="@+id/transaction_view_pager"
